move wishdir math, rename pathtracer methods
