+++ trunk/pathtracer/pathtracer.c
@@ -1,168 +1,12 @@
 #include "../quakedef.h"
 #include "pathtracer_private.h"
+#include "wishdir.h"
 
 extern cvar_t  show_speed_x;
 extern cvar_t  show_speed_y;
 
 static cvar_t scr_printbunnyhop = { "scr_printbunnyhop", "1" };
 static cvar_t scr_recordbunnyhop = { "scr_recordbunnyhop", "1" };
-
-// code from: https://github.com/shalrathy/quakespasm-shalrathy
-// Double checked the code, it is the same as in JoeQuake
-float Get_Wishdir_Speed_Delta(float angle) {
-	extern cvar_t sv_maxspeed;
-	extern cvar_t sv_friction;
-	extern cvar_t sv_edgefriction;
-	extern cvar_t sv_stopspeed;
-	extern cvar_t sv_accelerate;
-	extern usercmd_t cmd;
-	extern vec3_t velocitybeforethink;
-	extern qboolean onground;
-
-	int		i;
-	vec3_t		wishvel, wishdir, velocity, angles, origin, forward, right, up;
-	float		wishspeed, fmove, smove;
-
-	if (angle < -180) angle += 360;
-	if (angle > 180) angle -= 360;
-
-	for (i = 0;i < 3;i++) {
-		if (sv_player != NULL) {
-			angles[i] = sv_player->v.angles[i];
-		}
-		else {
-			angles[i] = 0;
-		}
-	}
-	angles[1] = angle;
-
-	AngleVectors(angles, forward, right, up);
-
-	fmove = cmd.forwardmove;
-	smove = cmd.sidemove;
-
-	// hack to not let you back into teleporter
-	if (sv_player != NULL && sv.time < sv_player->v.teleport_time && fmove < 0)
-		fmove = 0;
-
-	for (i = 0; i < 3; i++)
-		wishvel[i] = forward[i] * fmove + right[i] * smove;
-
-	if (sv_player != NULL && (int)sv_player->v.movetype != MOVETYPE_WALK)
-		wishvel[2] = cmd.upmove;
-	else
-		wishvel[2] = 0;
-
-	VectorCopy(wishvel, wishdir);
-	wishspeed = VectorNormalize(wishdir);
-	if (wishspeed > sv_maxspeed.value)
-	{
-		VectorScale(wishvel, sv_maxspeed.value / wishspeed, wishvel);
-		wishspeed = sv_maxspeed.value;
-	}
-
-	for (i = 0;i < 3;i++) velocity[i] = velocitybeforethink[i];
-	for (i = 0;i < 3;i++) {
-		if (sv_player != NULL) {
-			origin[i] = sv_player->v.origin[i];
-		}
-		else {
-			origin[i] = 0;
-		}
-	}
-
-	float oldspeed = sqrt(velocity[0] * velocity[0] + velocity[1] * velocity[1]);
-
-	if (sv_player != NULL && sv_player->v.movetype == MOVETYPE_NOCLIP)
-	{	// noclip
-		//VectorCopy (wishvel, velocity);
-	}
-	else if (onground)
-	{
-		while (1) { // SV_UserFriction ();
-			float* vel;
-			float	speed, newspeed, control;
-			vec3_t	start, stop;
-			float	friction;
-			trace_t	trace;
-
-			vel = velocity;
-
-			speed = sqrt(vel[0] * vel[0] + vel[1] * vel[1]);
-			if (!speed)
-				break;
-
-			// if the leading edge is over a dropoff, increase friction
-			start[0] = stop[0] = origin[0] + vel[0] / speed * 16;
-			start[1] = stop[1] = origin[1] + vel[1] / speed * 16;
-			start[2] = origin[2] + sv_player->v.mins[2];
-			stop[2] = start[2] - 34;
-
-			trace = SV_Move(start, vec3_origin, vec3_origin, stop, true, sv_player);
-
-			if (trace.fraction == 1.0)
-				friction = sv_friction.value * sv_edgefriction.value;
-			else
-				friction = sv_friction.value;
-
-			// apply friction
-			control = speed < sv_stopspeed.value ? sv_stopspeed.value : speed;
-			newspeed = speed - host_frametime * control * friction;
-
-			if (newspeed < 0)
-				newspeed = 0;
-			newspeed /= speed;
-
-			vel[0] = vel[0] * newspeed;
-			vel[1] = vel[1] * newspeed;
-			vel[2] = vel[2] * newspeed;
-			break;
-		}
-
-		while (1) { // SV_Accelerate (wishspeed, wishdir);
-			int			i;
-			float		addspeed, accelspeed, currentspeed;
-
-			currentspeed = DotProduct(velocity, wishdir);
-			addspeed = wishspeed - currentspeed;
-			if (addspeed <= 0)
-				break;
-			accelspeed = sv_accelerate.value * host_frametime * wishspeed;
-			if (accelspeed > addspeed)
-				accelspeed = addspeed;
-
-			for (i = 0; i < 3; i++)
-				velocity[i] += accelspeed * wishdir[i];
-			break;
-		}
-	}
-	else
-	{	// not on ground, so little effect on velocity
-		while (1) { // SV_AirAccelerate (wishspeed, wishvel);
-			int			i;
-			float		addspeed, wishspd, accelspeed, currentspeed;
-
-			wishspd = VectorNormalize(wishvel);
-			if (wishspd > 30)
-				wishspd = 30;
-			currentspeed = DotProduct(velocity, wishvel);
-			addspeed = wishspd - currentspeed;
-			if (addspeed <= 0)
-				break;
-			// accelspeed = sv_accelerate.value * host_frametime;
-			accelspeed = sv_accelerate.value * wishspeed * host_frametime;
-			if (accelspeed > addspeed)
-				accelspeed = addspeed;
-
-			for (i = 0; i < 3; i++)
-				velocity[i] += accelspeed * wishvel[i];
-			break;
-		}
-	}
-	float newspeed = sqrt(velocity[0] * velocity[0] + velocity[1] * velocity[1]);
-	return newspeed - oldspeed;
-}
-
 
 pathtracer_bunnyhop_t pathtracer_bunnyhop_samples[100];
 pathtracer_movement_t pathtracer_movement_samples[100000];
@@ -171,7 +15,7 @@ float drawbestangle = 0.f;
 extern	cvar_t	scr_printbunnyhop;
 extern	cvar_t	scr_recordbunnyhop;
 
-void R_DrawPathTracer(void)
+void PathTracer_Draw(void)
 {
 	if (sv_player == NULL) return;
 	if (!sv.active && !cls.demoplayback) return;
@@ -287,7 +131,7 @@ void R_DrawPathTracer(void)
 	glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
 }
 
-void SCR_DrawSpeedHelp(void) {
+void PathTracer_Sample(void) {
 
 	if (scr_printbunnyhop.value != 1.f) return;
 
@@ -337,7 +181,7 @@ void SCR_DrawSpeedHelp(void) {
 		onground0 = cl.onground;
 	}
 
-	// restarted
+	// Restart tracer
 	if (cl.time < prevcltime) {
 		memset(pathtracer_movement_samples, 0, sizeof(pathtracer_movement_samples));
 		pathtracer_movement_index = 0;
@@ -355,90 +199,92 @@ void SCR_DrawSpeedHelp(void) {
 		prevcltime = cl.time;
 	}
 
-	{
-		int angles = 90;
-		float bestspeed = 0.f;
-		float bestangle = 0;
-		bestangle = sv_player->v.angles[1];
-		bestspeed = Get_Wishdir_Speed_Delta(bestangle);
-		float playerspeed = Get_Wishdir_Speed_Delta(bestangle);
-		for (int i = 1; i < angles; i++) {
-			for (int neg = -1; neg <= 1; neg += 2) {
-				float curangle = 0.0f;
-				curangle = sv_player->v.angles[1] + i * neg;
-				float curspeed = Get_Wishdir_Speed_Delta(curangle);
-				if (curspeed > bestspeed) {
-					bestspeed = curspeed;
-					bestangle = curangle;
-				}
+		// Find best wishdir vector (largest speed increase in the plane, ígnores z)
+	int angles = 90;
+	float bestspeed = 0.f;
+	float bestangle = 0;
+	bestangle = sv_player->v.angles[1];
+	bestspeed = Get_Wishdir_Speed_Delta(bestangle);
+	float playerspeed = Get_Wishdir_Speed_Delta(bestangle);
+	for (int i = 1; i < angles; i++) {
+		for (int neg = -1; neg <= 1; neg += 2) {
+			float curangle = 0.0f;
+			curangle = sv_player->v.angles[1] + i * neg;
+			float curspeed = Get_Wishdir_Speed_Delta(curangle);
+			if (curspeed > bestspeed) {
+				bestspeed = curspeed;
+				bestangle = curangle;
 			}
 		}
-		for (int i = 1; i < 100; i++) {
-			for (int neg = -1; neg <= 1; neg += 2) {
-				float curangle = bestangle + neg * i / 100.0f;
-				float curspeed = Get_Wishdir_Speed_Delta(curangle);
-				if (curspeed > (bestspeed)) {
-					bestspeed = curspeed;
-					bestangle = curangle;
-				}
+	}
+	for (int i = 1; i < 100; i++) {
+		for (int neg = -1; neg <= 1; neg += 2) {
+			float curangle = bestangle + neg * i / 100.0f;
+			float curspeed = Get_Wishdir_Speed_Delta(curangle);
+			if (curspeed > (bestspeed)) {
+				bestspeed = curspeed;
+				bestangle = curangle;
 			}
 		}
-		if (!onground0) {
-			airspeedsum += Get_Wishdir_Speed_Delta(sv_player->v.angles[1]);
-			airspeedsummax += bestspeed;
-		}
+	}
+	if (!onground0) {
+		airspeedsum += Get_Wishdir_Speed_Delta(sv_player->v.angles[1]);
+		airspeedsummax += bestspeed;
+	}
 
-		if (bestangle < -180) bestangle += 360;
-		if (bestangle > 180) bestangle -= 360;
+	if (bestangle < -180) bestangle += 360;
+	if (bestangle > 180) bestangle -= 360;
 
 
-		float speed = sqrt(cl.velocity[0] * cl.velocity[0] + cl.velocity[1] * cl.velocity[1]);
+	float speed = sqrt(cl.velocity[0] * cl.velocity[0] + cl.velocity[1] * cl.velocity[1]);
 
-		boolean track = false;
-		if (pathtracer_movement_index > 0) {
-			pathtracer_movement_t* pms_prev = &pathtracer_movement_samples[pathtracer_movement_index - 1];
-			float dx = fabs(pms_prev->pos[0] - sv_player->v.origin[0]);
-			float dy = fabs(pms_prev->pos[1] - sv_player->v.origin[1]);
-			float dz = fabs(pms_prev->pos[2] - sv_player->v.origin[2]);
-			if (dx > .1f || dy > .1f || dz > .1f) {
-				track = true;
-			}
-		}
-		else {
+	boolean track = false;
+	if (pathtracer_movement_index > 0) {
+		pathtracer_movement_t* pms_prev = &pathtracer_movement_samples[pathtracer_movement_index - 1];
+		float dx = fabs(pms_prev->pos[0] - sv_player->v.origin[0]);
+		float dy = fabs(pms_prev->pos[1] - sv_player->v.origin[1]);
+		float dz = fabs(pms_prev->pos[2] - sv_player->v.origin[2]);
+		if (dx > .1f || dy > .1f || dz > .1f) {
 			track = true;
 		}
-		if (track && scr_recordbunnyhop.value == 1.f) {
-			pathtracer_movement_t* pms_new = &pathtracer_movement_samples[pathtracer_movement_index];
-			pms_new->pos[0] = sv_player->v.origin[0];
-			pms_new->pos[1] = sv_player->v.origin[1];
-			pms_new->pos[2] = sv_player->v.origin[2];
-			pms_new->velocity[0] = sv_player->v.velocity[0];
-			pms_new->velocity[1] = sv_player->v.velocity[1];
-			pms_new->velocity[2] = sv_player->v.velocity[2];
-			pms_new->angle = sv_player->v.angles[1];
+	}
+	else {
+		track = true;
+	}
+	// Sample movement
+	if (track && scr_recordbunnyhop.value == 1.f) {
+		pathtracer_movement_t* pms_new = &pathtracer_movement_samples[pathtracer_movement_index];
+		pms_new->pos[0] = sv_player->v.origin[0];
+		pms_new->pos[1] = sv_player->v.origin[1];
+		pms_new->pos[2] = sv_player->v.origin[2];
+		pms_new->velocity[0] = sv_player->v.velocity[0];
+		pms_new->velocity[1] = sv_player->v.velocity[1];
+		pms_new->velocity[2] = sv_player->v.velocity[2];
+		pms_new->angle = sv_player->v.angles[1];
 
-			if (ghost_entity.model != NULL) {
-				pms_new->pos[0] = ghost_entity.origin[0];
-				pms_new->pos[1] = ghost_entity.origin[1];
-				pms_new->pos[2] = ghost_entity.origin[2];
-				pms_new->velocity[0] = ghost_entity.origin[0] - ghost_entity.previousorigin[0];
-				pms_new->velocity[1] = ghost_entity.origin[1] - ghost_entity.previousorigin[1];
-				pms_new->velocity[2] = ghost_entity.origin[2] - ghost_entity.previousorigin[2];
-				pms_new->velocity[0] *= 200.f;
-				pms_new->velocity[1] *= 200.f;
-				pms_new->velocity[2] *= 200.f;
-				pms_new->angle = ghost_entity.angles[1];
-			}
-
-			pms_new->bestangle = bestangle;
-			pms_new->speed = speed;
-			pms_new->bestspeed = bestspeed;
-			pms_new->ongound = onground0;
-			pathtracer_movement_index++;
-			if (pathtracer_movement_index >= 100000) {
-				pathtracer_movement_index = 0;
-			}
+		if (ghost_entity.model != NULL) {
+			pms_new->pos[0] = ghost_entity.origin[0];
+			pms_new->pos[1] = ghost_entity.origin[1];
+			pms_new->pos[2] = ghost_entity.origin[2];
+			pms_new->velocity[0] = ghost_entity.origin[0] - ghost_entity.previousorigin[0];
+			pms_new->velocity[1] = ghost_entity.origin[1] - ghost_entity.previousorigin[1];
+			pms_new->velocity[2] = ghost_entity.origin[2] - ghost_entity.previousorigin[2];
+			pms_new->velocity[0] *= 200.f;
+			pms_new->velocity[1] *= 200.f;
+			pms_new->velocity[2] *= 200.f;
+			pms_new->angle = ghost_entity.angles[1];
 		}
+
+		pms_new->bestangle = bestangle;
+		pms_new->speed = speed;
+		pms_new->bestspeed = bestspeed;
+		pms_new->ongound = onground0;
+		pathtracer_movement_index++;
+		if (pathtracer_movement_index >= 100000) {
+			pathtracer_movement_index = 0;
+		}
+
+		// generate bunny hop hint
 		if ((onground0 && !prevonground) ||  // landed
 			(!onground0 && prevonground && prevprevonground)) { // take-off
 			char strextra[10];
@@ -496,6 +342,7 @@ void SCR_DrawSpeedHelp(void) {
 		prevonground = onground0;
 	}
 
+	// Find closest bunny hop hint
 	currentclosestbunnyhopDistance = 1000000000;
 	currentclosestbunnyhopIndex = -1;
 	for (int i = 0;i < 100;i++) {
@@ -519,7 +366,7 @@ void SCR_DrawSpeedHelp(void) {
 		}
 	}
 
-
+	// Draw closest bunny hop hint
 	if (currentclosestbunnyhopIndex >= 0 && currentclosestbunnyhopDistance < 20000.f) {
 		pathtracer_bunnyhop_samples[currentclosestbunnyhopIndex].selected = true;
