+++ trunk/pathtracer/wishdir.h
@@ -1,0 +1,8 @@
+#ifndef __WISHDIR_H
+#define __WISHDIR_H
+
+// code from: https://github.com/shalrathy/quakespasm-shalrathy
+// Double checked the code, it is the same as in JoeQuake
+float Get_Wishdir_Speed_Delta(float angle);
+
+#endif // __WISHDIR_H
